refactor(staking): split DelegationPortfolioStore into multiple files
