+++ packages/staking/src/features/store/delegationPortfolioStore/mapStakePoolToDisplayData.ts
@@ -32,7 +32,9 @@ export const mapStakePoolToDisplayData = ({
     margin,
     name: stakePool.metadata?.name || '-',
     owners: stakePool.owners ? stakePool.owners.map((owner: Wallet.Cardano.RewardAccount) => owner.toString()) : [],
-    pledge: stakePool.pledge ? `${Wallet.util.lovelacesToAdaString(stakePool.pledge.toString())}${cardanoCoinSymbol}` : '-',
+    pledge: stakePool.pledge
+      ? `${Wallet.util.lovelacesToAdaString(stakePool.pledge.toString())}${cardanoCoinSymbol}`
+      : '-',
     retired: stakePool.status === Wallet.Cardano.StakePoolStatus.Retired,
     saturation: stakePool.metrics?.saturation && formatPercentages(stakePool.metrics.saturation),
     size: `${stakePool.metrics?.size.live ?? '-'} %`,
