+++ packages/staking/src/features/store/delegationPortfolioStore/stateMachine/index.ts
@@ -1,0 +1,12 @@
+export { processExpandedViewCases } from './processExpandedViewCases';
+export { processPopupViewCases } from './processPopupViewCases';
+export type { Command } from './commands';
+export { Flow, DrawerDefaultStep, DrawerManagementStep } from './types';
+export type {
+  CurrentPortfolioStakePool,
+  DelegationPortfolioState,
+  DraftPortfolioStakePool,
+  DrawerStep,
+  ExecuteCommand,
+  Handler,
+} from './types';
