+++ packages/staking/src/features/store/delegationPortfolioStore/stateMachine/stateTreeUtilities.ts
@@ -1,0 +1,18 @@
+import { Command } from './commands';
+import { Handler } from './types';
+
+export const cases =
+  <T extends string>(definition: Record<T, Handler>, discriminator: T, parentName: string): Handler =>
+  (params) => {
+    const handler = definition[discriminator];
+    if (!handler) {
+      console.error(`Invalid discriminator ${discriminator} in ${parentName} handlers`);
+      return;
+    }
+    handler(params);
+  };
+
+export const handler =
+  <C extends Command>(handlerBody: Handler<C>): Handler<C> =>
+  (params) =>
+    handlerBody(params);
