+++ packages/staking/src/features/store/delegationPortfolioStore/stateMachine/types.ts
@@ -1,0 +1,66 @@
+import { Wallet } from '@lace/cardano';
+import { AdaSymbol } from '../types';
+import { Command } from './commands';
+
+export type StakePoolWithLogo = Wallet.Cardano.StakePool & { logo?: string };
+
+export type DraftPortfolioStakePool = Wallet.Cardano.Cip17Pool & {
+  displayData: Wallet.util.StakePool;
+};
+
+export type CurrentPortfolioStakePool = DraftPortfolioStakePool & {
+  displayData: Wallet.util.StakePool & {
+    lastReward: bigint;
+    totalRewards: bigint;
+  };
+  stakePool: Wallet.Cardano.StakePool;
+  value: bigint;
+};
+
+export enum Flow {
+  Overview = 'Overview',
+  BrowsePools = 'BrowsePools',
+  CurrentPoolDetails = 'CurrentPoolDetails',
+  PoolDetails = 'PoolDetails',
+  PortfolioManagement = 'PortfolioManagement',
+  ChangingPreferences = 'ChangingPreferences',
+  NewPortfolio = 'NewPortfolio',
+}
+
+export type ExpandedViewFlow = Flow;
+
+export type PopupViewFLow = Flow.Overview | Flow.CurrentPoolDetails;
+
+export enum DrawerDefaultStep {
+  PoolDetails = 'PoolDetails',
+}
+
+export enum DrawerManagementStep {
+  Preferences = 'Preferences',
+  Confirmation = 'Confirmation',
+  Sign = 'Sign',
+  Success = 'Success',
+  Failure = 'Failure',
+}
+
+export type DrawerStep = DrawerDefaultStep | DrawerManagementStep;
+
+export type DelegationPortfolioState = {
+  activeDrawerStep?: DrawerStep;
+  activeFlow: Flow;
+  cardanoCoinSymbol: AdaSymbol;
+  pendingSelectedPortfolio?: DraftPortfolioStakePool[];
+  currentPortfolio: CurrentPortfolioStakePool[];
+  draftPortfolio?: DraftPortfolioStakePool[];
+  selectedPortfolio: DraftPortfolioStakePool[];
+  view?: 'popup' | 'expanded';
+  viewedStakePool?: StakePoolWithLogo;
+};
+
+export type ExecuteCommand = <C extends Command>(command: C) => void;
+
+export type Handler<C extends Command = any> = (params: {
+  command: C;
+  executeCommand: ExecuteCommand;
+  state: DelegationPortfolioState;
+}) => void;
