+++ packages/staking/src/features/store/delegationPortfolioStore/useDelegationPortfolioStore.ts
@@ -1,0 +1,122 @@
+import { Wallet } from '@lace/cardano';
+import { create } from 'zustand';
+import { immer } from 'zustand/middleware/immer';
+import { CARDANO_COIN_SYMBOL, LAST_STABLE_EPOCH, targetWeight } from './constants';
+import { mapStakePoolToDisplayData } from './mapStakePoolToDisplayData';
+import {
+  Command,
+  DelegationPortfolioState,
+  ExecuteCommand,
+  Flow,
+  Handler,
+  processExpandedViewCases,
+  processPopupViewCases,
+} from './stateMachine';
+import { DelegationPortfolioStore } from './types';
+
+const defaultState: DelegationPortfolioState = {
+  activeDrawerStep: undefined,
+  activeFlow: Flow.Overview,
+  cardanoCoinSymbol: 'ADA',
+  currentPortfolio: [],
+  draftPortfolio: undefined,
+  pendingSelectedPortfolio: undefined,
+  selectedPortfolio: [],
+  view: undefined,
+  viewedStakePool: undefined,
+};
+
+export const useDelegationPortfolioStore = create(
+  immer<DelegationPortfolioStore>((set, get) => ({
+    ...defaultState,
+    mutators: {
+      executeCommand: (command) => {
+        let numberOfRecursiveCalls = 0;
+        const callsConsideredAnInfiniteLoop = 10;
+        let paramsStack: Command[] = [command];
+
+        const { view } = get();
+        // eslint-disable-next-line unicorn/consistent-function-scoping
+        let processCommand: Handler = () => {
+          throw new Error('DelegationPortfolioStore: view not set');
+        };
+        if (view === 'popup') {
+          processCommand = processPopupViewCases;
+        }
+        if (view === 'expanded') {
+          processCommand = processExpandedViewCases;
+        }
+
+        set((state) => {
+          // TODO: decide whether to throw this function away
+          //  as we may not want to execute command from inside the SM
+          const executeCommand: ExecuteCommand = (childCommand) => {
+            paramsStack = [...paramsStack, childCommand];
+            numberOfRecursiveCalls += 1;
+            if (numberOfRecursiveCalls > callsConsideredAnInfiniteLoop) {
+              const error = new Error('DelegationPortfolioStore: Infinite loop detected');
+              throw Object.assign(error, { paramsStack });
+            }
+            // eslint-disable-next-line sonarjs/no-extra-arguments
+            processCommand({
+              command: childCommand,
+              executeCommand,
+              state,
+            });
+          };
+
+          // eslint-disable-next-line sonarjs/no-extra-arguments
+          processCommand({ command, executeCommand, state });
+        });
+      },
+      forceAbortFlows: () =>
+        set((state) => {
+          const viewingOverviewPage = [Flow.Overview, Flow.CurrentPoolDetails, Flow.PortfolioManagement].includes(
+            state.activeFlow
+          );
+          state.activeFlow = viewingOverviewPage ? Flow.Overview : Flow.BrowsePools;
+          state.activeDrawerStep = undefined;
+          state.selectedPortfolio = [];
+          state.pendingSelectedPortfolio = undefined;
+          state.viewedStakePool = undefined;
+        }),
+      setCardanoCoinSymbol: (currentChain) =>
+        set((state) => {
+          state.cardanoCoinSymbol = CARDANO_COIN_SYMBOL[currentChain.networkId];
+        }),
+      setCurrentPortfolio: async ({ currentEpoch, delegationDistribution, delegationRewardsHistory }) => {
+        const lastNonVolatileEpoch = currentEpoch.epochNo.valueOf() - LAST_STABLE_EPOCH;
+        const confirmedRewardHistory = delegationRewardsHistory.all.filter(
+          ({ epoch }) => epoch.valueOf() <= lastNonVolatileEpoch
+        );
+        const currentPortfolio = delegationDistribution.map(({ pool: stakePool, percentage, stake }) => {
+          const confirmedPoolRewards = confirmedRewardHistory
+            .filter(({ poolId }) => poolId === stakePool.id)
+            .map(({ rewards }) => rewards);
+
+          return {
+            displayData: {
+              ...mapStakePoolToDisplayData({ cardanoCoinSymbol: get().cardanoCoinSymbol, stakePool }),
+              lastReward: confirmedPoolRewards[confirmedPoolRewards.length - 1] || BigInt(0),
+              totalRewards: Wallet.BigIntMath.sum(confirmedPoolRewards),
+            },
+            id: stakePool.hexId,
+            name: stakePool.metadata?.name,
+            stakePool,
+            ticker: stakePool.metadata?.ticker,
+            value: stake,
+            weight: Math.round(percentage * targetWeight),
+          };
+        });
+
+        set((state) => {
+          state.currentPortfolio = currentPortfolio;
+        });
+      },
+      setView: (view) =>
+        set((state) => {
+          state.view = view;
+        }),
+    },
+  }))
+);
