+++ packages/staking/src/features/store/index.ts
@@ -1,2 +1,2 @@
 export * from './stakingStore';
-export * from './useDelegationPortfolioStore';
+export * from './delegationPortfolioStore';
