--- packages/staking/src/features/store/types.ts
@@ -1,1 +1,0 @@
-export type AdaSymbol = 'ADA' | 'tADA';
